refactor map controls into their own component
