+++ src/interface/src/app/app.module.ts
@@ -17,6 +17,7 @@ import { AppComponent } from './app.component';
 import { FeaturesModule } from './features/features.module';
 import { LoginComponent } from './login/login.component';
 import { LayerInfoCardComponent } from './map/layer-info-card/layer-info-card.component';
+import { MapControlPanelComponent } from './map/map-control-panel/map-control-panel.component';
 import { MapNameplateComponent } from './map/map-nameplate/map-nameplate.component';
 import { MapComponent } from './map/map.component';
 import { PlanCreateDialogComponent } from './map/plan-create-dialog/plan-create-dialog.component';
@@ -53,6 +54,7 @@ import { TopBarComponent } from './top-bar/top-bar.component';
     MapNameplateComponent,
     PlanCreateDialogComponent,
     LayerInfoCardComponent,
+    MapControlPanelComponent,
   ],
   imports: [
     AppRoutingModule,
