+++ src/interface/src/app/map/map-control-panel/map-control-panel.component.scss
@@ -1,0 +1,118 @@
+.controls-container {
+  background-color: white;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  padding-top: 20px;
+  width: 400px;
+}
+
+.legend-container {
+  padding: 0px 24px 0px 24px;
+}
+
+// Styling for the buttons to switch between 1, 2, and 4 maps
+.map-count-button-row {
+  display: flex;
+  justify-content: center;
+}
+
+.map-count-button {
+  align-items: center;
+  background-color: #f1eff2;
+  border-radius: 8px;
+  display: flex;
+  flex-direction: column;
+  height: 28px;
+  justify-content: center;
+  margin-left: 2px;
+  margin-right: 2px;
+  width: 72px;
+}
+
+.map-count-button.selected {
+  background-color: #e4e2e5;
+}
+
+.map-count-button-grid-row {
+  display: flex;
+  justify-content: center;
+}
+
+.map-count-button-grid-cell {
+  border: 1px solid #656367;
+  border-radius: 1px;
+  box-sizing: border-box;
+  height: 8px;
+  margin-right: 1px;
+  margin-bottom: 1px;
+  width: 8px;
+}
+
+.layer-controls-divider {
+  margin-bottom: 20px;
+  margin-top: 20px;
+}
+
+.layer-control-container {
+  display: flex;
+  flex-direction: row;
+}
+
+.layer-controls-tab-group {
+  flex: 1;
+  overflow-y: auto;
+}
+
+.layer-radio-group {
+  align-items: flex-start;
+  display: flex;
+  flex-direction: column;
+  margin: 15px 0;
+}
+
+.layer-radio-button {
+  margin: 5px;
+}
+
+.layer-checkbox {
+  margin: 0px 5px 0px;
+}
+
+.condition-tree-invisible {
+  display: none;
+}
+
+.condition-tree ul,
+.condition-tree li {
+  margin-top: 0;
+  margin-bottom: 0;
+  list-style-type: none;
+}
+
+/*
+ * This padding sets alignment of the nested nodes.
+ */
+.condition-tree .mat-nested-tree-node div[role=group] {
+  padding-left: 20px;
+}
+
+/*
+ * Padding for leaf nodes.
+ * Leaf nodes need to have padding so as to align with other non-leaf nodes
+ * under the same parent.
+ */
+.condition-tree div[role=group] > .mat-tree-node {
+  padding-left: 20px;
+}
+
+.mat-tree-node > .disabled {
+  opacity: 0.5;
+}
+
+// Workaround for unresolved Angular bug re: tab width on desktop
+// (https://github.com/angular/components/issues/2829).
+::ng-deep .mat-tab-label {
+  min-width: 60px !important;
+}
