+++ src/interface/src/app/map/map.component.ts
@@ -1,4 +1,3 @@
-import { NestedTreeControl } from '@angular/cdk/tree';
 import {
   AfterViewInit,
   ApplicationRef,
@@ -10,7 +9,6 @@ import {
 } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { MatTreeNestedDataSource } from '@angular/material/tree';
 import { Router } from '@angular/router';
 import { Feature, Geometry } from 'geojson';
 import {
@@ -32,11 +30,9 @@ import {
   SessionService,
 } from '../services';
 import {
-  BaseLayerType,
   BoundaryConfig,
   colormapConfigToLegend,
   ConditionsConfig,
-  DataLayerConfig,
   DEFAULT_COLORMAP,
   defaultMapConfig,
   defaultMapViewOptions,
@@ -46,7 +42,6 @@ import {
   MapViewOptions,
   NONE_BOUNDARY_CONFIG,
   NONE_COLORMAP,
-  NONE_DATA_LAYER_CONFIG,
   Region,
 } from '../types';
 import { MapManager } from './map-manager';
@@ -57,14 +52,6 @@ export enum AreaCreationAction {
   NONE = 0,
   DRAW = 1,
   UPLOAD = 2,
-}
-
-export interface ConditionsNode extends DataLayerConfig {
-  showInfoIcon?: boolean;
-  infoMenuOpen?: boolean;
-  disableSelect?: boolean; // Node should not include a radio button
-  styleDisabled?: boolean; // Node should be greyed out but still selectable
-  children?: ConditionsNode[];
 }
 
 @Component({
@@ -88,15 +75,6 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
   selectedRegion$: Observable<Region | null>;
   planState$: Observable<PlanState>;
   selectedMap$: Observable<Map | undefined>;
-
-  readonly noneBoundaryConfig = NONE_BOUNDARY_CONFIG;
-
-  readonly baseLayerTypes: number[] = [
-    BaseLayerType.Road,
-    BaseLayerType.Terrain,
-    BaseLayerType.Satellite,
-  ];
-  readonly BaseLayerType = BaseLayerType;
 
   existingProjectsGeoJson$ = new BehaviorSubject<GeoJSON.GeoJSON | null>(null);
 
@@ -133,11 +111,6 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
   showAreaCreationActionButtons = false;
   selectedAreaCreationAction: AreaCreationAction = AreaCreationAction.NONE;
   showConfirmAreaButton$ = new BehaviorSubject(false);
-
-  conditionTreeControl = new NestedTreeControl<ConditionsNode>(
-    (node) => node.children
-  );
-  conditionDataSource = new MatTreeNestedDataSource<ConditionsNode>();
 
   private readonly destroy$ = new Subject<void>();
 
@@ -199,19 +172,6 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
     this.mapManager.polygonsCreated$
       .pipe(takeUntil(this.destroy$))
       .subscribe(this.showConfirmAreaButton$);
-
-    this.conditionDataSource.data = [NONE_DATA_LAYER_CONFIG];
-    this.conditionsConfig$
-      .pipe(filter((config) => !!config))
-      .subscribe((config) => {
-        this.conditionDataSource.data = this.conditionsConfigToData(config!);
-        this.maps.forEach((map) => {
-          // Ensure the radio button corresponding to the saved selection is selected.
-          map.config.dataLayerConfig = this.findAndRevealNodeWithFilepath(
-            map.config.dataLayerConfig.filepath
-          );
-        });
-      });
   }
 
   ngOnInit(): void {
@@ -686,145 +646,5 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
         }
         return '0%';
     }
-  }
-
-  /** Gets the legend that should be shown in the sidebar.
-   *
-   *  WARNING: This function is run constantly and shouldn't do any heavy lifting!
-   */
-  getSelectedLegend(): Legend | undefined {
-    return this.maps[this.mapViewOptions$.getValue().selectedMapIndex].legend;
-  }
-
-  /** Used to compute whether a node in the condition layer tree has children. */
-  hasChild = (_: number, node: ConditionsNode) =>
-    !!node.children && node.children.length > 0;
-
-  /** Used to compute whether to show the info button on a condition layer node. */
-  showInfoIcon = (node: ConditionsNode) =>
-    node.filepath?.length && (node.showInfoIcon || node.infoMenuOpen);
-
-  /** Unstyles all nodes in the tree using recursion. */
-  unstyleAllNodes(node?: ConditionsNode): void {
-    if (!node && this.conditionDataSource.data.length > 0) {
-      this.conditionDataSource.data.forEach((node) =>
-        this.unstyleAllNodes(node)
-      );
-    } else if (node) {
-      node.styleDisabled = false;
-      node.children?.forEach((child) => this.unstyleAllNodes(child));
-    }
-  }
-
-  /** Visually indicates that all the descendants of a condition layer node are
-   *  included in the current analysis by setting their style, using recursion.
-   */
-  styleDescendantsDisabled(node: ConditionsNode): void {
-    node.children?.forEach((child) => {
-      child.styleDisabled = true;
-      this.styleDescendantsDisabled(child);
-    });
-  }
-
-  onSelect(node: ConditionsNode): void {
-    this.unstyleAllNodes();
-    this.styleDescendantsDisabled(node);
-  }
-
-  private conditionsConfigToData(config: ConditionsConfig): ConditionsNode[] {
-    return [
-      NONE_DATA_LAYER_CONFIG,
-      {
-        ...config,
-        display_name: 'Current condition',
-        disableSelect: true,
-        children: config.pillars
-          ?.filter((pillar) => pillar.display)
-          .map((pillar): ConditionsNode => {
-            return {
-              ...pillar,
-              disableSelect: true,
-              children: pillar.elements?.map((element): ConditionsNode => {
-                return {
-                  ...element,
-                  disableSelect: true,
-                  children: element.metrics,
-                };
-              }),
-            };
-          }),
-      },
-      {
-        display_name: 'Current condition (normalized)',
-        filepath: config.filepath?.concat('_normalized'),
-        colormap: config.colormap,
-        normalized: true,
-        disableSelect: true,
-        children: config.pillars
-          ?.filter((pillar) => pillar.display)
-          .map((pillar): ConditionsNode => {
-            return {
-              ...pillar,
-              filepath: pillar.filepath?.concat('_normalized'),
-              normalized: true,
-              children: pillar.elements?.map((element): ConditionsNode => {
-                return {
-                  ...element,
-                  filepath: element.filepath?.concat('_normalized'),
-                  normalized: true,
-                  children: element.metrics?.map((metric): ConditionsNode => {
-                    return {
-                      ...metric,
-                      filepath: metric.filepath?.concat('_normalized'),
-                      normalized: true,
-                    };
-                  }),
-                };
-              }),
-            };
-          }),
-      },
-    ];
-  }
-
-  /** Find the node matching the given filepath in the condition tree (if any), and expand its ancestors
-   *  so it becomes visible.
-   */
-  private findAndRevealNodeWithFilepath(
-    filepath: string | undefined
-  ): ConditionsNode {
-    if (!filepath || filepath === NONE_DATA_LAYER_CONFIG.filepath)
-      return NONE_DATA_LAYER_CONFIG;
-    for (let tree of this.conditionDataSource.data) {
-      if (tree.filepath === filepath) return tree;
-      if (tree.children) {
-        for (let pillar of tree.children) {
-          if (pillar.filepath === filepath) {
-            this.conditionTreeControl.expand(tree);
-            return pillar;
-          }
-          if (pillar.children) {
-            for (let element of pillar.children) {
-              if (element.filepath === filepath) {
-                this.conditionTreeControl.expand(tree);
-                this.conditionTreeControl.expand(pillar);
-                return element;
-              }
-              if (element.children) {
-                for (let metric of element.children) {
-                  if (metric.filepath === filepath) {
-                    this.conditionTreeControl.expand(tree);
-                    this.conditionTreeControl.expand(pillar);
-                    this.conditionTreeControl.expand(element);
-                    return metric;
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    return NONE_DATA_LAYER_CONFIG;
   }
 }
